Provides a middleware for parsing HTTP body
At this time, this middleware, ELOQUENT.MVC.MIDDLEWARE:PARSE-BODY just
supports parsing HTTP body when the CONTENT-TYPE is
"application/json". If this matched, the middleware will extract the
HTTP body as a string, and parse this string by JSON syntax. Finally,
the parsing result will set into the :BODY key in the EXTRAS hash-table
in request