A macro for binding variables from JSON body
The macro ELOQUENT.MVC.CONTROLLER:JSON-BODY-BIND is for this. It's
similar to the macro QUERY-STRING-BIND under the same package. It's
strongly depends on the middleware PARSE-BODY, because it assumes the
object parsed from JSON HTTP body is put under the :BODY key in the
EXTRAS slot in a request instance