Pass URL parameters to action directly
By specifying call convention in router rules, the action handler is
able to get rid of a parameter for instance of class
ELOQUENT.MVC.REQUEST:<REQUEST>. It's able to focus on the business
logic, and this action can be tested in REPL because there's no need to
construct a HTTP request for calling it. This mechanism is exprimental yet.