Support passing parameters from query string
Specify the keyword argument :QUERY-STRING-BIND in a router rule, the
dispatcher is able to pass values extracted from query string to the
action handler. All these values are structured as a plist and appended
to the URL parameters, and finally passed to action handler as keyword
arguments. This keyword specifier is not equivalent to the macro
QUERY-STRING-BIND provided by package ELOQUENT.MVC.CONTROLLER, because
it's unable to do type cast or provide default value yet.