Define router system as middleware
The procedures of finding a rule and calling an action is extracted from
the code of dispatcher. The routing logics is defined as the middleware
ELOQUENT.MVC.MIDDLEWARE:SET-MATCHED-RULE, and the logics of invoking an
action is defined as APPLY-MATCHED-RULE. Now the middlewares are able to
know which rule the current request matched. And the logics like
NOT-FOUND could be implemented as a middleware, and STATIC-FILE could be
implemented as a regular action function