Provides ability for base64 DECODE/ENCODE
The new package ELOQUENT.MVC.PRELUDE.BASE64 wraps the APIs exported
from library CL-BASE64. And the DECODE and ENCODE functions from this
new package change the semantics of keyword parameter URI. If URI is
not NIL, the equal, plus and slash signs in resulting string will be
replaced by empty, minus and underscore signs.

This new package aims to be used for implementing JWT generation.