Reimplements STATIC-FILE as a helper
The original version of STATIC-FILE is a middleware provided by
package ELOQUENT.MVC.CONTRIB, it handles the client requests start
with a prefix specified in configuration file. Since the machenism of
calling the middlewares and actions changed, the STATIC-FILE can no
longer be implemented as a middleware. Therefor, it's defined as a
helper function, which should be mentioned in file router.lisp.